tasks.py refactoring + front-end fix
