+++ templates/index.html
@@ -77,16 +77,6 @@
                                 <label for="config-clustering_runs">Clustering Runs:</label>
                                 <input type="number" id="config-clustering_runs">
                             </div>
-                             <div id="kmeans-params-basic" style="display: grid; gap: 1rem;">
-                                <div>
-                                    <label for="config-num_clusters_min">Min Clusters:</label>
-                                    <input type="number" id="config-num_clusters_min">
-                                </div>
-                                <div>
-                                    <label for="config-num_clusters_max">Max Clusters:</label>
-                                    <input type="number" id="config-num_clusters_max">
-                                </div>
-                            </div>
                             <div class="advanced-param hidden">
                                 <label for="config-max_distance">Max Distance:</label>
                                 <input type="number" step="0.01" id="config-max_distance">
@@ -141,6 +131,21 @@
                             </div>
                         </div>
 
+                        <!-- K-Means specific parameters (for Basic and Advanced K-Means view) -->
+                        <div id="kmeans-params-basic" class="hidden"> <!-- Initially hidden; JS controls visibility. Not an 'advanced-param' itself. -->
+                            <h4>K-Means Specific:</h4>
+                            <div class="param-group"> <!-- param-group for styling the inputs -->
+                                <div>
+                                    <label for="config-num_clusters_min">Min Clusters:</label>
+                                    <input type="number" id="config-num_clusters_min">
+                                </div>
+                                <div>
+                                    <label for="config-num_clusters_max">Max Clusters:</label>
+                                    <input type="number" id="config-num_clusters_max">
+                                </div>
+                            </div>
+                        </div>
+
                         <div id="dbscan-params" class="advanced-param hidden">
                             <h4>DBSCAN Specific:</h4>
                             <div>
@@ -159,10 +164,6 @@
                                 <label for="config-dbscan_min_samples_max">DBSCAN Min Samples Max:</label>
                                 <input type="number" id="config-dbscan_min_samples_max">
                             </div>
-                        </div>
-
-                        <div id="kmeans-params" class="advanced-param hidden">
-                           <h4>K-Means Specific:</h4>
                         </div>
 
                         <div id="gmm-params" class="advanced-param hidden">
